[refactor] Restructure access requester integration for Slack and Teams
- Introduced `requester_context.py` to centralize Boto3 client management and configuration for both Slack and Teams integrations.
- Refactored `main.py` to utilize the new requester context, reducing complexity and improving maintainability.
- Created `slack_app.py` to encapsulate Slack-specific functionality, ensuring a clear separation of concerns.
- Updated `revoker.py` to streamline notifier initialization based on the chat platform.
- Enhanced Teams message handling in `teams_handlers.py` for improved command processing.
- Cleaned up unused imports and optimized existing functions for better performance.

This refactor enhances the overall architecture, making it easier to manage and extend the integration with both Slack and Teams.
