+++ src/slack_app.py
@@ -1,0 +1,506 @@
+"""Slack Bolt app for the access requester — import only when ``chat_platform`` is ``slack``."""
+
+from __future__ import annotations
+
+import uuid
+from datetime import timedelta
+from typing import Callable
+
+from slack_bolt import Ack, App, BoltContext
+from slack_sdk import WebClient
+from slack_sdk.web.slack_response import SlackResponse
+
+import access_control
+import config
+import entities
+import group
+import organizations
+import request_store
+import schedule
+import slack_helpers
+import sso
+from entities.elevator_request import ElevatorRequestKind, ElevatorRequestRecord, ElevatorRequestStatus
+from errors import SSOUserNotFound, handle_errors
+from requester_context import RequesterContext
+
+logger = config.get_logger(service="slack_app")
+
+_slack_app: App | None = None
+
+
+def get_slack_app(ctx: RequesterContext) -> App:
+    """Build and cache the Slack :class:`App` once per cold start."""
+    global _slack_app
+    if _slack_app is None:
+        _slack_app = _build_slack_app(ctx)
+    return _slack_app
+
+
+def _build_slack_app(ctx: RequesterContext) -> App:
+    cfg = ctx.cfg
+    schedule_client = ctx.schedule_client
+    org_client = ctx.org_client
+    sso_client = ctx.sso_client
+    identity_store_client = ctx.identity_store_client
+    s3_client = ctx.s3_client
+
+    def build_initial_form_handler(
+        view_class: slack_helpers.RequestForAccessView | slack_helpers.RequestForGroupAccessView,
+    ) -> Callable[[WebClient, dict, Ack], SlackResponse]:
+        def show_initial_form_for_request(
+            client: WebClient,
+            body: dict,
+            ack: Ack,
+        ) -> SlackResponse:
+            ack()
+            if view_class == slack_helpers.RequestForGroupAccessView and not cfg.group_statements:
+                return client.chat_postMessage(
+                    channel=cfg.slack_channel_id,
+                    text="Group statements are not configured, please check the configuration. Or use another /command.",
+                )
+            if view_class == slack_helpers.RequestForAccessView and not cfg.statements:
+                return client.chat_postMessage(
+                    channel=cfg.slack_channel_id,
+                    text="Statements are not configured, please check the configuration. Or use another /command.",
+                )
+
+            # Try getting SSO user to check if user exist
+            try:
+                sso.get_user_principal_id_by_email(
+                    identity_store_client=identity_store_client,
+                    identity_store_id=sso.describe_sso_instance(sso_client, cfg.sso_instance_arn).identity_store_id,
+                    email=slack_helpers.get_user(client, id=body.get("user", {}).get("id")).email,
+                    cfg=cfg,
+                )
+
+            except SSOUserNotFound:
+                client.chat_postMessage(
+                    channel=cfg.slack_channel_id,
+                    text=f"<@{body.get('user', {}).get('id') or 'UNKNOWN_USER'}>,"
+                    "Your request for AWS permissions failed because SSO Elevator could not find your user in SSO."
+                    "This often happens if your AWS SSO email differs from your Slack email."
+                    "Please check the SSO Elevator logs for more details.",
+                )
+                raise
+
+            logger.info(f"Showing initial form for {view_class.__name__}")
+            logger.debug("Request body", extra={"body": body})
+            trigger_id = body["trigger_id"]
+            user_id = body.get("user", {}).get("id")
+            callback_id = view_class.CALLBACK_ID
+
+            response = client.views_open(trigger_id=trigger_id, view=view_class.build())
+            if user_id:
+                request_store.put_view_id(
+                    str(user_id),
+                    callback_id,
+                    str(response.data["view"]["id"]),  # type: ignore # noqa: PGH003
+                )
+            logger.debug("Stored view_id in request store for modal update")
+
+            return response
+
+        return show_initial_form_for_request
+
+    def load_select_options_for_group_access_request(client: WebClient, body: dict) -> SlackResponse:
+        logger.info("Loading select options for view (groups)")
+        logger.debug("Request body", extra={"body": body})
+        sso_instance = sso.describe_sso_instance(sso_client, cfg.sso_instance_arn)
+        groups = sso.get_groups_from_config(sso_instance.identity_store_id, identity_store_client, cfg)
+
+        user_id = body.get("user", {}).get("id")
+        callback_id = slack_helpers.RequestForGroupAccessView.CALLBACK_ID
+
+        view_id = request_store.get_view_id(str(user_id), callback_id) if user_id else None
+        if not view_id:
+            logger.warning(
+                f"View ID not found for user {user_id} callback {callback_id}. "
+                "This happens when Lambda container is recycled between shortcut invocations. "
+                "Opening a new view as fallback."
+            )
+            # Fallback: open a new view with the data already loaded
+            trigger_id = body["trigger_id"]
+            view = slack_helpers.RequestForGroupAccessView.update_with_groups(groups=groups)
+            return client.views_open(trigger_id=trigger_id, view=view)
+
+        view = slack_helpers.RequestForGroupAccessView.update_with_groups(groups=groups)
+        return client.views_update(view_id=view_id, view=view)
+
+    def load_select_options_for_account_access_request(client: WebClient, body: dict) -> SlackResponse:
+        logger.info("Loading select options for view (accounts and permission sets)")
+        logger.debug("Request body", extra={"body": body})
+
+        accounts = organizations.get_accounts_from_config_with_cache(org_client=org_client, s3_client=s3_client, cfg=cfg)
+        permission_sets = sso.get_permission_sets_from_config_with_cache(sso_client=sso_client, s3_client=s3_client, cfg=cfg)
+
+        user_id = body.get("user", {}).get("id")
+        callback_id = slack_helpers.RequestForAccessView.CALLBACK_ID
+
+        view_id = request_store.get_view_id(str(user_id), callback_id) if user_id else None
+        if not view_id:
+            logger.warning(
+                f"View ID not found for user {user_id} callback {callback_id}. "
+                "This happens when Lambda container is recycled between shortcut invocations. "
+                "Opening a new view as fallback."
+            )
+            # Fallback: open a new view with the data already loaded
+            trigger_id = body["trigger_id"]
+            view = slack_helpers.RequestForAccessView.update_with_accounts_and_permission_sets(
+                accounts=accounts, permission_sets=permission_sets
+            )
+            return client.views_open(trigger_id=trigger_id, view=view)
+
+        view = slack_helpers.RequestForAccessView.update_with_accounts_and_permission_sets(accounts=accounts, permission_sets=permission_sets)
+        return client.views_update(view_id=view_id, view=view)
+
+    app = App(
+        process_before_response=True,
+        # Logger removed to avoid pickle errors with lazy listeners in Lambda
+        # Slack Bolt will use its own default logger instead
+    )
+
+    app.shortcut("request_for_access")(
+        build_initial_form_handler(view_class=slack_helpers.RequestForAccessView),  # type: ignore # noqa: PGH003
+        load_select_options_for_account_access_request,
+    )
+
+    app.shortcut("request_for_group_membership")(
+        build_initial_form_handler(view_class=slack_helpers.RequestForGroupAccessView),  # type: ignore # noqa: PGH003
+        load_select_options_for_group_access_request,
+    )
+
+    @handle_errors
+    def handle_button_click(body: dict, client: WebClient, context: BoltContext) -> SlackResponse:  # noqa: ARG001
+        logger.info("Handling button click")
+        try:
+            payload = slack_helpers.ButtonClickedPayload.model_validate(body)
+        except Exception as e:
+            logger.exception(e)
+            return group.handle_group_button_click(body, client, context)
+
+        logger.info("Button click payload", extra={"payload": payload})
+        # Approver might be from different Slack workspace, if so, get_user will fail.
+        try:
+            approver = slack_helpers.get_user(client, id=payload.approver_slack_id)
+        except Exception as e:
+            logger.warning(f"Failed to get approver user info: {e}")
+            return client.chat_postMessage(
+                channel=payload.channel_id,
+                text=f"""Unable to process this approval - approver information could not be retrieved.
+            This may happen if the approver <@{payload.approver_slack_id}> is from a different Slack workspace.
+            Please check the module configuration.""",
+                thread_ts=payload.thread_ts,
+            )
+        requester = slack_helpers.get_user(client, id=payload.request.requester_slack_id)
+        is_user_in_channel = slack_helpers.check_if_user_is_in_channel(client, cfg.slack_channel_id, requester.id)
+
+        if not request_store.try_begin_in_flight_approval(
+            requester_slack_id=payload.request.requester_slack_id,
+            account_id=payload.request.account_id,
+            permission_set_name=payload.request.permission_set_name,
+            group_id=None,
+        ):
+            return client.chat_postMessage(
+                channel=payload.channel_id,
+                text=f"<@{approver.id}> request is already in progress, please wait for the result.",
+                thread_ts=payload.thread_ts,
+            )
+
+        if payload.action == entities.ApproverAction.Discard:
+            blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+                blocks=payload.message["blocks"],
+                color_coding_emoji=cfg.bad_result_emoji,
+            )
+
+            blocks = slack_helpers.remove_blocks(blocks, block_ids=["buttons"])
+            blocks.append(slack_helpers.button_click_info_block(payload.action, approver.id).to_dict())
+
+            text = f"Request was discarded by<@{approver.id}> "
+            dm_text = f"Your request was discarded by <@{approver.id}>."
+            client.chat_update(
+                channel=payload.channel_id,
+                ts=payload.thread_ts,
+                blocks=blocks,
+                text=text,
+            )
+            if payload.elevator_request_id:
+                request_store.update_request_status(payload.elevator_request_id, ElevatorRequestStatus.discarded)
+
+            request_store.end_in_flight_approval(
+                requester_slack_id=payload.request.requester_slack_id,
+                account_id=payload.request.account_id,
+                permission_set_name=payload.request.permission_set_name,
+                group_id=None,
+            )
+            if cfg.send_dm_if_user_not_in_channel and not is_user_in_channel:
+                logger.info(f"User {requester.id} is not in the channel. Sending DM with message: {dm_text}")
+                client.chat_postMessage(channel=requester.id, text=dm_text)
+            return client.chat_postMessage(
+                channel=payload.channel_id,
+                text=text,
+                thread_ts=payload.thread_ts,
+            )
+
+        decision = access_control.make_decision_on_approve_request(
+            action=payload.action,
+            statements=cfg.statements,
+            account_id=payload.request.account_id,
+            permission_set_name=payload.request.permission_set_name,
+            approver_email=approver.email,
+            requester_email=requester.email,
+        )
+        logger.info("Decision on request was made", extra={"decision": decision.dict()})
+
+        if not decision.permit:
+            request_store.end_in_flight_approval(
+                requester_slack_id=payload.request.requester_slack_id,
+                account_id=payload.request.account_id,
+                permission_set_name=payload.request.permission_set_name,
+                group_id=None,
+            )
+            return client.chat_postMessage(
+                channel=payload.channel_id,
+                text=f"<@{approver.id}> you can not approve this request",
+                thread_ts=payload.thread_ts,
+            )
+
+        text = f"Permissions granted to <@{requester.id}> by <@{approver.id}>."
+        dm_text = f"Your request was approved by <@{approver.id}>. Permissions granted."
+        blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+            blocks=payload.message["blocks"],
+            color_coding_emoji=cfg.good_result_emoji,
+        )
+
+        blocks = slack_helpers.remove_blocks(blocks, block_ids=["buttons"])
+        blocks.append(slack_helpers.button_click_info_block(payload.action, approver.id).to_dict())
+        is_user_in_channel = slack_helpers.check_if_user_is_in_channel(client, cfg.slack_channel_id, requester.id)
+        client.chat_update(
+            channel=payload.channel_id,
+            ts=payload.thread_ts,
+            blocks=blocks,
+            text=text,
+        )
+
+        access_control.execute_decision(
+            decision=decision,
+            permission_set_name=payload.request.permission_set_name,
+            account_id=payload.request.account_id,
+            permission_duration=payload.request.permission_duration,
+            approver=approver,
+            requester=requester,
+            reason=payload.request.reason,
+            elevator_request_id=payload.elevator_request_id,
+        )
+        if payload.elevator_request_id:
+            request_store.update_request_status(payload.elevator_request_id, ElevatorRequestStatus.completed)
+        request_store.end_in_flight_approval(
+            requester_slack_id=payload.request.requester_slack_id,
+            account_id=payload.request.account_id,
+            permission_set_name=payload.request.permission_set_name,
+            group_id=None,
+        )
+        if cfg.send_dm_if_user_not_in_channel and not is_user_in_channel:
+            logger.info(f"User {requester.id} is not in the channel. Sending DM with message: {dm_text}")
+            client.chat_postMessage(channel=requester.id, text=dm_text)
+        return client.chat_postMessage(
+            channel=payload.channel_id,
+            text=text,
+            thread_ts=payload.thread_ts,
+        )
+
+    def acknowledge_request(ack: Ack) -> None:
+        ack()
+
+    app.action(entities.ApproverAction.Approve.value)(
+        ack=acknowledge_request,
+        lazy=[handle_button_click],
+    )
+
+    app.action(entities.ApproverAction.Discard.value)(
+        ack=acknowledge_request,
+        lazy=[handle_button_click],
+    )
+
+    @handle_errors
+    def handle_request_for_access_submittion(  # noqa: PLR0915, PLR0912
+        body: dict,
+        ack: Ack,  # noqa: ARG001
+        client: WebClient,
+        context: BoltContext,  # noqa: ARG001
+    ) -> SlackResponse | None:
+        logger.info("Handling request for access submission")
+        request = slack_helpers.RequestForAccessView.parse(body)
+        logger.info("View submitted", extra={"view": request})
+        requester = slack_helpers.get_user(client, id=request.requester_slack_id)
+        decision = access_control.make_decision_on_access_request(
+            cfg.statements,
+            account_id=request.account_id,
+            permission_set_name=request.permission_set_name,
+            requester_email=requester.email,
+        )
+        logger.info("Decision on request was made", extra={"decision": decision.dict()})
+
+        account = organizations.describe_account(org_client, request.account_id)
+
+        elevator_id = str(uuid.uuid4())
+        request_store.put_access_request(
+            ElevatorRequestRecord(
+                elevator_request_id=elevator_id,
+                kind=ElevatorRequestKind.account,
+                status=ElevatorRequestStatus.awaiting_approval,
+                requester_slack_id=request.requester_slack_id,
+                reason=request.reason,
+                permission_duration_seconds=int(request.permission_duration.total_seconds()),
+                account_id=request.account_id,
+                permission_set_name=request.permission_set_name,
+            )
+        )
+
+        show_buttons = bool(decision.approvers)
+        slack_response = client.chat_postMessage(
+            blocks=slack_helpers.build_approval_request_message_blocks(
+                sso_client=sso_client,
+                identity_store_client=identity_store_client,
+                slack_client=client,
+                requester_slack_id=request.requester_slack_id,
+                account=account,
+                role_name=request.permission_set_name,
+                reason=request.reason,
+                permission_duration=request.permission_duration,
+                show_buttons=show_buttons,
+                color_coding_emoji=cfg.waiting_result_emoji,
+                elevator_request_id=elevator_id,
+            ),
+            channel=cfg.slack_channel_id,
+            text=f"Request for access to {account.name} account from {requester.real_name}",
+        )
+        if slack_response.get("ts") is not None:
+            request_store.update_slack_presentation(elevator_id, cfg.slack_channel_id, str(slack_response["ts"]))
+
+        if show_buttons:
+            ts = slack_response["ts"]
+            if ts is not None:
+                schedule.schedule_discard_buttons_event(
+                    schedule_client=schedule_client,
+                    time_stamp=ts,
+                    channel_id=cfg.slack_channel_id,
+                    elevator_request_id=elevator_id,
+                )
+                schedule.schedule_approver_notification_event(
+                    schedule_client=schedule_client,
+                    message_ts=ts,
+                    channel_id=cfg.slack_channel_id,
+                    time_to_wait=timedelta(
+                        minutes=cfg.approver_renotification_initial_wait_time,
+                    ),
+                    elevator_request_id=elevator_id,
+                )
+
+        match decision.reason:
+            case access_control.DecisionReason.ApprovalNotRequired:
+                text = "Approval for this Permission Set & Account is not required. Request will be approved automatically."
+                dm_text = "Approval for this Permission Set & Account is not required. Your request will be approved automatically."
+                color_coding_emoji = cfg.good_result_emoji
+            case access_control.DecisionReason.SelfApproval:
+                text = "Self approval is allowed and requester is an approver. Request will be approved automatically."
+                dm_text = "Self approval is allowed and you are an approver. Your request will be approved automatically."
+                color_coding_emoji = cfg.good_result_emoji
+            case access_control.DecisionReason.RequiresApproval:
+                approvers, approver_emails_not_found = slack_helpers.find_approvers_in_slack(
+                    client,
+                    decision.approvers,  # type: ignore # noqa: PGH003
+                )
+                if not approvers:
+                    text = """
+                None of the approvers from configuration could be found in Slack.
+                Request cannot be processed. Please discard the request and check the module configuration.
+                """
+                    dm_text = """
+                Your request cannot be processed because none of the approvers from configuration could be found in Slack.
+                Please discard the request and check the module configuration.
+                """
+                    color_coding_emoji = cfg.bad_result_emoji
+                else:
+                    mention_approvers = " ".join(f"<@{approver.id}>" for approver in approvers)
+                    text = f"{mention_approvers} there is a request waiting for the approval."
+                    if approver_emails_not_found:
+                        missing_emails = ", ".join(approver_emails_not_found)
+                        text += f"""
+                    Note: Some approvers ({missing_emails}) could not be found in Slack.
+                    Please discard the request and check the module configuration.
+                    """
+                    dm_text = f"Your request is waiting for the approval from {mention_approvers}."
+                    color_coding_emoji = cfg.waiting_result_emoji
+            case access_control.DecisionReason.NoApprovers:
+                text = "Nobody can approve this request."
+                dm_text = "Nobody can approve this request."
+                color_coding_emoji = cfg.bad_result_emoji
+            case access_control.DecisionReason.NoStatements:
+                text = "There are no statements for this Permission Set & Account."
+                dm_text = "There are no statements for this Permission Set & Account."
+                color_coding_emoji = cfg.bad_result_emoji
+
+        is_user_in_channel = slack_helpers.check_if_user_is_in_channel(client, cfg.slack_channel_id, requester.id)
+
+        logger.info(f"Sending message to the channel {cfg.slack_channel_id}, message: {text}")
+        client.chat_postMessage(text=text, thread_ts=slack_response["ts"], channel=cfg.slack_channel_id)
+        if cfg.send_dm_if_user_not_in_channel and not is_user_in_channel:
+            logger.info(f"User {requester.id} is not in the channel. Sending DM with message: {dm_text}")
+            client.chat_postMessage(
+                channel=requester.id,
+                text=f"""
+            {dm_text} You are receiving this message in a DM because you are not a member of the channel <#{cfg.slack_channel_id}>.
+            """,
+            )
+
+        blocks = slack_helpers.HeaderSectionBlock.set_color_coding(
+            blocks=slack_response["message"]["blocks"],
+            color_coding_emoji=color_coding_emoji,
+        )
+        client.chat_update(
+            channel=cfg.slack_channel_id,
+            ts=slack_response["ts"],
+            blocks=blocks,
+            text=text,
+        )
+
+        access_control.execute_decision(
+            decision=decision,
+            permission_set_name=request.permission_set_name,
+            account_id=request.account_id,
+            permission_duration=request.permission_duration,
+            approver=requester,
+            requester=requester,
+            reason=request.reason,
+            elevator_request_id=elevator_id,
+        )
+        if decision.grant and elevator_id:
+            request_store.update_request_status(elevator_id, ElevatorRequestStatus.completed)
+
+        if decision.grant:
+            client.chat_postMessage(
+                channel=cfg.slack_channel_id,
+                text=f"Permissions granted to <@{requester.id}>",
+                thread_ts=slack_response["ts"],
+            )
+            if not is_user_in_channel and cfg.send_dm_if_user_not_in_channel:
+                client.chat_postMessage(
+                    channel=requester.id,
+                    text="Your request was processed, permissions granted.",
+                )
+
+    app.view(slack_helpers.RequestForAccessView.CALLBACK_ID)(
+        ack=acknowledge_request,
+        lazy=[handle_request_for_access_submittion],
+    )
+
+    app.view(slack_helpers.RequestForGroupAccessView.CALLBACK_ID)(
+        ack=acknowledge_request,
+        lazy=[group.handle_request_for_group_access_submittion],
+    )
+
+    @app.action("duration_picker_action")
+    def handle_duration_picker_action(ack: Ack) -> None:
+        ack()
+
+    return app
